fix_categories_products_conflict
