+++ lib/features/product/providers/supplier_categories_provider.dart
@@ -2,9 +2,10 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:tzwad_mobile/core/file_upload/picked_file_controller.dart';
 import 'package:tzwad_mobile/core/network/api_service_provider.dart';
+import 'package:tzwad_mobile/features/category/ui/categories_supplier/category_supplier_controller.dart';
+import 'package:tzwad_mobile/features/category/ui/categories_supplier/category_supplier_state.dart';
 import 'package:tzwad_mobile/features/product/repository/product_supplier_repository.dart';
-import 'package:tzwad_mobile/features/product/ui/products_supplier/product_supplier_controller.dart';
-import 'package:tzwad_mobile/features/product/ui/products_supplier/product_supplier_state.dart';
+
 
 
 final supplierSearchQueryProvider = StateProvider<String>((ref) => '');
@@ -12,18 +13,19 @@ final supplierSearchQueryProvider = StateProvider<String>((ref) => '');
 final isSearchVisibleProvider = StateProvider<bool>((ref) => false);
 
 
-final productSupplierRepositoryProvider = Provider.autoDispose<ProductSupplierRepository>((ref) {
+final categorySupplierRepositoryProvider = Provider.autoDispose<CategorySupplierRepository>((ref) {
     final apiService = ref.read(
       apiServiceProvider,
     );
-    return ProductSupplierRepository(apiService: apiService,);
+    return CategorySupplierRepository(apiService: apiService,);
   },
 );
 
 
-final productSupplierControllerProvider = NotifierProvider.autoDispose<ProductSupplierController, ProductSupplierState>(
+final categorySupplierControllerProvider = NotifierProvider.autoDispose<CategorySupplierController,
+    CategorySupplierState>(
       () {
-    return ProductSupplierController();
+    return CategorySupplierController();
   },
 );
